Enforce semi-colons in js and add them where missing
There is no "right" answer to the question of whether semi-colons
are good or bad in modern-day js, but historically Mozilla's
web tools (and in particular our data tools) have used and enforced
a style where semicolons are always used to terminate the end of a line
or expression.

Redash, test-tube, metricsgraphics, missioncontrol for example all
use this rule. Furthermore, it's a default part of the airbnb
style guide which is recommended by Mozilla's frontend infrastructure
working group (https://github.com/mozilla-frontend-infra).

To avoid confusing people who have to move between these projects on a
frequent basis, I propose that we just adopt the rule and move on.
This commit adds the rule (really deleting an exception) and updates
the source code accordingly (done automatically via eslint's --fix
option)