Persist session across page reloads
Do away with storing the oauth token locally, and inject it down into a
templated copy of the iodide notebook that is served by Django. Eventually
we probably want to do away with this approach entirely, but this at least
makes saving to gist a little bit more convenient for now.