In order to make this project more accessible for new contributors, I went ahaed and documented the files found here.
Since I am not yet an Iodide expert, please forgive me if I made any
mistakes, but every newbie needs to go down this path, so let us
make it easier for all of them.

There are also a few questions in this document, for things
I was not sure about.