Temporary fork markdown-it katex plugin
The markdown-it plugin for katex hasn't been updated since 2016, so
the katex version was out of date. By temporarily forking the plugin,
we can use the latest katex package, which includes lots of fixes.

This commit should be reversed as soon as the markdown-it katex plugin
is actively maintained again.