Do not block compilation for prettier warnings
It can be a little frustrating when prettier blocks compilation during
development. This often happens to me while tinkering with props in JSX.
Adding or removing a prop can put the line over or under the 80
character limit, making prettier require whitespace changes. Until those
stylistic changes are made, the new code cannot be compiled or tested in
the browser.