Clear local autosave on load if no different + fix local autosave detection
It is occasionally possible for us to have a copy of a local autosave
which has identical contents to what's already on the server in certain
rare cases (e.g. if we send a no-op update to the server).

We actually need a better model of what's synced/not synced with the
server to handle this properly, but I believe this fix will handle 99% of
the likely cases.