Fix notebook creation for users who do not yet exist (#2234)
We should allow users with correct privileges to create notebooks on behalf of
users that do not yet exist. This fixes an issue we had with our "explore query
in iodide" redash plugin, which failed for users who had not logged into iodide
before.

This change also slightly changes the error condition when "forking" a notebook
when it is does not exist (400 -> 404).