Move "clear declared variables" to workspace pane
After doing this, it is a bit unclear to me whether we should keep this
functionality or just remove it altogether -- it's a bit of a footgun,
doesn't work as advertised (e.g. #2101) and doesn't seem that useful even
on its own terms (I have never actually tried to use it). But in any case,
this seems like a more logical place for the action than the menu.