Client side for /from-template/ view
Since a POST request initiated from another domain is unauthenticated,
check to see if we're logged in in this case, and auto login
and create the notebook for this case.

As part of this work, make sure that we *really* delete all cookies
and JWT tokens when we log out.